Moving tests to separate files
