+++ test/conftest.py
@@ -1,1 +1,1 @@
-pytest_plugins = ["test.helpers.auth", "test.helpers.database", "test.helpers.requests"]
+pytest_plugins = ["test.helpers.auth", "test.helpers.database", "test.helpers.request"]
